# 8273322: Enhance macro logic optimization for masked logic operations.