--- a/src/hotspot/cpu/x86/c2_MacroAssembler_x86.cpp
+++ b/src/hotspot/cpu/x86/c2_MacroAssembler_x86.cpp
@@ -4155,20 +4155,20 @@ void C2_MacroAssembler::vector_castF2I_evex(XMMRegister dst, XMMRegister src, XM
 void C2_MacroAssembler::evpternlog(XMMRegister dst, int func, KRegister mask, XMMRegister src2, XMMRegister src3,
                                    bool merge, BasicType bt, int vlen_enc) {
   if (bt == T_INT) {
-    evpternlogd(dst, func, mask, src2, src3, true, vlen_enc);
+    evpternlogd(dst, func, mask, src2, src3, merge, vlen_enc);
   } else {
     assert(bt == T_LONG, "");
-    evpternlogq(dst, func, mask, src2, src3, true, vlen_enc);
+    evpternlogq(dst, func, mask, src2, src3, merge, vlen_enc);
   }
 }
 
 void C2_MacroAssembler::evpternlog(XMMRegister dst, int func, KRegister mask, XMMRegister src2, Address src3,
                                    bool merge, BasicType bt, int vlen_enc) {
   if (bt == T_INT) {
-    evpternlogd(dst, func, mask, src2, src3, true, vlen_enc);
+    evpternlogd(dst, func, mask, src2, src3, merge, vlen_enc);
   } else {
     assert(bt == T_LONG, "");
-    evpternlogq(dst, func, mask, src2, src3, true, vlen_enc);
+    evpternlogq(dst, func, mask, src2, src3, merge, vlen_enc);
   }
 }