--- a/src/hotspot/cpu/x86/x86.ad
+++ b/src/hotspot/cpu/x86/x86.ad
@@ -9575,10 +9575,10 @@ instruct vternlog_reg_masked(vec dst, vec src2, vec src3, immU8 func, kReg mask)
   match(Set dst (MacroLogicV dst (Binary src2 (Binary src3 (Binary func mask)))));
   format %{ "vternlog_masked $dst,$src2,$src3,$func,$mask\t! vternlog masked operation" %}
   ins_encode %{
-    int vector_len = vector_length_encoding(this);
+    int vlen_enc = vector_length_encoding(this);
     BasicType bt = Matcher::vector_element_basic_type(this);
     __ evpternlog($dst$$XMMRegister, $func$$constant, $mask$$KRegister,
-                  $src2$$XMMRegister, $src3$$XMMRegister, true, bt, vector_len);
+                  $src2$$XMMRegister, $src3$$XMMRegister, true, bt, vlen_enc);
   %}
   ins_pipe( pipe_slow );
 %}
@@ -9587,10 +9587,10 @@ instruct vternlogd_mem_masked(vec dst, vec src2, memory src3, immU8 func, kReg m
   match(Set dst (MacroLogicV dst (Binary src2 (Binary src3 (Binary func mask)))));
   format %{ "vternlog_masked $dst,$src2,$src3,$func,$mask\t! vternlog masked operation" %}
   ins_encode %{
-    int vector_len = vector_length_encoding(this);
+    int vlen_enc = vector_length_encoding(this);
     BasicType bt = Matcher::vector_element_basic_type(this);
     __ evpternlog($dst$$XMMRegister, $func$$constant, $mask$$KRegister,
-                  $src2$$XMMRegister, $src3$$Address, true, bt, vector_len);
+                  $src2$$XMMRegister, $src3$$Address, true, bt, vlen_enc);
   %}
   ins_pipe( pipe_slow );
 %}