@vnkozlov 2022-01-05
`compiler/vectorapi/TestMaskedMacroLogicVector.java` test failed on Aarch64 machines:
```
Unrecognized VM option 'UseAVX=3'
Error: Could not create the Java Virtual Machine.
Error: A fatal exception has occurred. Program will exit.

java.lang.RuntimeException: TestFramework flag VM exited with 1
	at compiler.lib.ir_framework.driver.FlagVMProcess.checkFlagVMExitCode(FlagVMProcess.java:135)
	at compiler.lib.ir_framework.driver.FlagVMProcess.start(FlagVMProcess.java:121)
	at compiler.lib.ir_framework.driver.FlagVMProcess.<init>(FlagVMProcess.java:63)
	at compiler.lib.ir_framework.TestFramework.start(TestFramework.java:658)
	at compiler.lib.ir_framework.TestFramework.start(TestFramework.java:322)
	at compiler.lib.ir_framework.TestFramework.runWithFlags(TestFramework.java:230)
	at compiler.vectorapi.TestMaskedMacroLogicVector.main(TestMaskedMacroLogicVector.java:837)
```