@jatin-bhateja 2022-01-05
> I think whole "Bitwise operation packing optimization" code should be moved out from `compile.cpp`. May be to `vectornode.cpp where `MacroLogicVNode` code is located.
> 
Hi @vnkozlov ,
Yes we can also extended AndV/OrV/XorV/AndVMask/OrVMask/XorVMask idealizations to perform macro logic folding, 
current changes keeps the implementation clean and limited to one optimization stage.

> Copyright year should be updated to 2022 in all changed files.